Moved Combat related files to a Combat folder
Moved the board scripts and the combat manager to a new folder called
'Combat'. No files were deleted, just moved.
