+++ Assets/Scripts/Combat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d08d135a724b7d74484a12626340910d
+folderAsset: yes
+timeCreated: 1489519879
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
