+++ Assets/Scripts/Combat/Board/Information/GridResizeInformation.cs.meta
