+++ Assets/Scripts/Combat/Board/Information/MatchInformation.cs
@@ -7,11 +7,9 @@
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace Board.Information
+namespace Combat.Board.Information
 {
     using System.Collections.Generic;
-
-    using UnityEngine;
 
     public class MatchInformation
     {
