+++ Assets/Scripts/Combat/CombatLifeBar.cs
@@ -1,0 +1,32 @@
+﻿namespace Combat
+{
+    using UnityEngine;
+    using UnityEngine.UI;
+
+    [RequireComponent(typeof(Image))]
+    public class CombatLifeBar : MonoBehaviour
+    {
+        [SerializeField]
+        private Image m_Image;
+        [SerializeField]
+        private Text m_HealthText;
+
+        // Use this for initialization
+        private void Awake()
+        {
+            m_Image = GetComponent<Image>();
+            OnCombatModeChange();
+
+            m_HealthText = GetComponentInChildren<Text>();
+            m_HealthText.text =
+                GameManager.self.playerData.health.totalValue + "/" + GameManager.self.playerData.health.value;
+
+            CombatManager.self.onCombatModeChange.AddListener(OnCombatModeChange);
+        }
+
+        private void OnCombatModeChange()
+        {
+            m_Image.color = CombatManager.self.combatUiInformation.currentModeUiInformation.modeColor;
+        }
+    }
+}
