+++ Assets/Scripts/Combat/CombatLifeBar.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ca345e9c09ca22a4d8de3a7b068a21bc
+timeCreated: 1489519765
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
